requirements.txt
openai
python-dotenv
gunicorn